Source Code/requirements.txt
numpy==1.16.3
torch==1.0.1
spacy==2.1.3
transformers==2.2.2
Cython==0.29.10
tqdm==4.32.2
neuralcoref==4.0
argparse
scikit-learn
pytest
deepspeech
rogue
pydub
scipy
librosa unidecode inflect librosa